back/requirements.txt
alembic==1.13.0
amqp==5.2.0
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
billiard==4.2.0
boto3==1.34.20
botocore==1.34.20
celery==5.3.6
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
dnspython==2.6.1
email_validator==2.1.1
exceptiongroup==1.1.3
fastapi==0.104.1
filelock==3.13.1
fonttools==4.46.0
fsspec==2023.12.2
greenlet==3.0.1
h11==0.14.0
idna==3.4
importlib-resources==6.1.1
Jinja2==3.1.2
jmespath==1.0.1
kafka-python==2.0.2
kiwisolver==1.4.5
kombu==5.3.5
Mako==1.3.0
MarkupSafe==2.1.3
matplotlib==3.8.2
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.2
onnx==1.15.0
opencv-python==4.8.1.78
packaging==23.2
pandas==2.1.4
passlib==1.7.4
Pillow==10.1.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.0
psycopg2-binary==2.9.9
py-cpuinfo==9.0.0
pydantic==1.10.7
pydantic_core==2.14.3
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.4
requests==2.31.0
s3transfer==0.10.0
scipy==1.11.4
seaborn==0.13.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
sympy==1.12
thop==0.1.1.post2209072238
# torch==2.1.2
torchvision==0.16.2
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
ultralytics==8.0.227
urllib3==2.0.7
uvicorn==0.24.0.post1
vine==5.1.0
wcwidth==0.2.13
zipp==3.17.0
ffmpeg-python==0.2.0
pytorch-ignite==0.5.2